Refactor std functions
Moved to another top module to reduce the size “instrinsics/mod.rs”.
