+++ src/dyon_std/lifetimechk.rs
