+++ src/lib.rs
@@ -30,6 +30,7 @@ pub mod vec4;
 pub mod write;
 
 mod grab;
+mod dyon_std;
 
 pub use runtime::Runtime;
 pub use prelude::{Lt, Prelude, Dfn};
